Rename manageGeneSets modal component
